// frontend/src/api/models/handlerSupplierResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerSupplierResponseStatus } from './handlerSupplierResponseStatus'

/**
 * Supplier details returned by the API
 */
export interface HandlerSupplierResponse {
  address?: string
  attributes?: string
  bank_account?: string
  bank_account_name?: string
  bank_name?: string
  city?: string
  code?: string
  contact_name?: string
  country?: string
  created_at?: string
  credit_limit?: number
  email?: string
  full_address?: string
  id?: string
  name?: string
  notes?: string
  payment_term_days?: number
  phone?: string
  postal_code?: string
  province?: string
  rating?: number
  short_name?: string
  sort_order?: number
  status?: HandlerSupplierResponseStatus
  tax_id?: string
  tenant_id?: string
  updated_at?: string
  version?: number
}