// frontend/src/api/models/handlerSupplierResponseStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerSupplierResponseStatus =
  (typeof HandlerSupplierResponseStatus)[keyof typeof HandlerSupplierResponseStatus]

export const HandlerSupplierResponseStatus = {
  active: 'active',
  inactive: 'inactive',
  blocked: 'blocked',
} as const